Move ContextReader to its own file
